+++ src/routes.js
@@ -8,7 +8,7 @@ import { loadOrganization } from './actions/navActions.js'
 import Home from './containers/home'
 import SignPetition from './containers/sign-petition'
 import ThanksPage from './pages/thanks'
-import SearchPage from './pages/search'
+import SearchPage from './containers/search'
 import PetitionCreatorDashboard from './containers/petition-creator-dashboard'
 import CreatePetitionPage from './pages/create-petition-page'
 import PetitionReportPage from './pages/petition-report-page'
